dbTransaction implementation
